requirements.txt
# File: requirements.txt
# FE-AI System Dependencies

# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-plotly>=0.0.1

# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0

# Federated Learning
flwr>=1.5.0
flower>=1.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Signal Processing
librosa>=0.10.0
pywavelets>=1.4.1
spectrum>=0.8.1

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
bokeh>=3.2.0

# Explainable AI
shap>=0.42.0
lime>=0.2.0.1

# Database
pymongo>=4.5.0
motor>=3.3.0
mongoengine>=0.27.0

# Model Deployment
onnx>=1.14.0
onnxruntime>=1.16.0
joblib>=1.3.0

# File Processing
openpyxl>=3.1.0
xlrd>=2.0.1
python-multipart>=0.0.6

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Monitoring & Logging
wandb>=0.16.0
mlflow>=2.7.0
loguru>=0.7.0

# Security & Privacy
cryptography>=41.0.0
pycryptodome>=3.19.0
hashlib-compat>=1.0.1

# Utilities
tqdm>=4.66.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Performance
numba>=0.58.0
cupy-cuda11x>=12.0.0  # Optional: for GPU acceleration

# Web Scraping (for dataset collection)
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Time Series
statsmodels>=0.14.0
pmdarima>=2.0.0

# Jupyter Support
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0